Abstracted EncodedParams class into a separate file
Added nodoc to the class since it is already documented in `http_service.rb`
Tested back to ruby-2.2.3
